Rename the repo and crates
The current names are long and annoying, lets rename them.

This is based on top of the 0.3.0 releases so that there is some
continuity.

The repository changes from `rust-bitcoind-json-rpc` to `corepc`.

Crate name changes:

- `bitcoind-json-rpc-client` becomes `corepc-client`
- `bitcoind-json-rpc-types` becomes `corepc-types`
- `bitcoind-json-rpc-regtest` becomes `corepc-node`

Directory name changes:

- `json` becomes `types` (to match the crate name)
- `regtest` becomes `bitcoind` (to match the original project name)

Next, bump version numbers

In preparation for release bump the version number, add changelog
entries, and update the lock-files.

Done for all three released crates.