Add verify tool
Add a tool to help verify that what we have done is correct.

Adds crate that when run produces:

```bash
Verifying support for Bitcoin Core v17

        Checking he documented methods are correct for this version of Core ... ✓
        The documented status is correct ... ✓
```

Run the tool for v17 module and fix the few mistakes it found - WIN!

Then fix up the docs in v18 and run the tool to verify the v18 stuff
is correct.