Reduce number of supported minor versions
We support Core back to v0.17.1 and currently explicitly test all minor
versions back to v22 - this is probably an overkill and is only going to
increase as Core progresses. Instead lets just test all minor versions
of the last three releases and then the latest minor version for older
releases.

This has the benefit of reducing the number of CI jobs.