verify: Verify tests using test output file
Currently the verify integration test code is broken because it greps
for test names not bothering that some may be feature gated.

Instead of grepping for integration tests by name instead accept an
option that includes a file that is the output from running the
integration tests and then grep that file.