Remove feature gating from test_p2p unit test
This test works for `28_0`. It does not need feature gating out.

The reason this was done is because it cannot be run in parallel with
`test_multi_p2p`. We can leave that one feature guarded out in v28
builds.

The underlying issue is a problem with how Bitcoin Core binds to ports
and is tracked in #41