Enable running multiple Core v28 nodes
Currently the `node` crate cannot run multiple instances of Bitcoin Core
v28 because of changes to bind logic in that release.

The solution is to use `-bind` instead of `-port`.

For more information on the relevant changes in Core v28 see:

https://github.com/bitcoin/bitcoin/blob/master/doc/release-notes/release-notes-28.0.md#p2p-and-network-changes

This patch was created with help of @MaxFangX - mad props.

Fix: #41