Remove default feature
The default feature in node overides an explicit version feature. This
is a footgun. Currently the logic around using `bitcoind` from the
environment is borked.

Also, features in the `node` crate are unusual in that at least one
version feature is expected. Lets make this a build requirement.

Maintain the current behaviour in relation to `--all-features` i.e., all
features implies latest version 28.0

- Remove `default` feature
- Remove `download` from version number features
- Use `compile_error` to catch missing version feature