Include node in workspace
Currently we do not include the `node` crate in the workspace because
`run_task` did not support testing it.

However `run_task` has now been patched.

Include `node` in the workspace. Configure it to work with the newly
patched `run_task`. Add a `extra_tests.sh` script that runs the `node`
tests.

In order to get MSRV building we make all deps use `default-features =
false` then explicitly enable just what we need.

Also change the feature guarding in the build script so that
`--all-features` works. This is already supported in the main crate code
since the highest version feature overrides the lower ones.