Include node in workspace
Currently we do not include the `node` crate in the workspace because
`run_task` did not support testing it.

However `run_task` is patched in

 https://github.com/rust-bitcoin/rust-bitcoin-maintainer-tools/pull/19

So we can use it now.

Include `node` in the workspace. Configure it to work with the newly
patched `run_task`. Add a `extra_tests.sh` script that runs the `node`
tests.

In order to get MSRV building we make all deps use `default-features =
false` then explicitly enable just what we need.