verify: Update VERSIONS
The `VERSIONS` const is an array that is supposed to include all
supported versions but I forgot to update it as I added support to the
tool for newer versions.

FTR the const is used in the output of `verify v99` (i.e. unsupported
version) and also in `verify all`.