types: Remove duplicate rpc api files
With the recent updates to the `verify` tool we put the output of
`bitcoin-cli help` into the `verify` crate. But I forgot to remove the
ones it `types`, we new have duplicate copies.

Elect to leave the ones in `verify` since they are parsed by that
tool during verification of the repository.