node: Set default feature to 0_17_1
In #45 we removed the default feature because we were setting it to
`28_0` and this was a footgun because downstream crates needed to
explicitly use no default features. This change made it impossible to
build the crate without selecting a version feature which implicitly
means `--no-default-features` does not work.

Instead we can set the default to `0_17_1` and then the crate builds
with no default features and also downstream users who select a higher
version get correct behaviour because higher versions override lower
ones.