Add missing transaction categories
Contrary to Bitcoin Core's v17 docs, there are *5* transaction
categories that might be returned as part of a response to
`listtransactions`, `listsinceblock`, or `gettransaction`. While Core
fixed this omission in the docs in https://github.com/bitcoin/bitcoin/pull/14653,
we didn't account for them in the respective model's `enum`, leading to
calls to `get_transaction` randomly failing with the rather obscure
error message:
```
Err(JsonRpc(Json(Error("unknown variant `generate`, expected `send` or `receive`", line: 1, column: 682))))
```

Here, we fix this omission and add the missing categories.