node: Fix the rustdocs build
Currently the docs.rs build is not working. From the log it can be
seen that the build is trying to download a bitcoind binary. I cannot
see why this is happening, it does not happen locally. Anyway we can
just change the docs build to use `28_0` instead of all features and
this bypasses the problem.

While we are at it, change the feature gating in `build.rs` on the
`download` module because it still seems wrong. Phew logically
combining negative truth values is hard to reason about.