types: Update status for getblockstats
Recently we added tests for `getblockstats` but forgot to update the
status.

There is still some confusion in my mind as what should be done about
versions 19-24. Currently the method is untested and not working for
these versions but we still re-export the `GetBlockStats` type and the
method exists in `bitcoincli help getblockstats`.