Test getblockstats for all versions
I was confused by `getblockstats` not working and thought it did not
work for version 19-24 however there was a bug in a helper function that
meant we always tried to get block stats for the genesis block and this
does not work for these versions. That bug has been fixed, now we can
test and set the status of `getblockstats` for all versions.