Add deprecated `warning` field for `loadwallet/`cratewallet` responses
With `v25` the `warning` field was deprecated for `warnings`. While it
should only be available when the `-deprecatedrpc=walletwarningfield` is
passed, I was hitting parse errors even though I did not find where I
passed this argument. Here, we add the `warning` field to the `v25`
types (although we ignore it as it's literally duplicative of what's in
`warnings` already) to avoid the parse error.