Fix up unloadwallet
Starting at Core version 21 the `unloadwallet` method starts returning a
json object (with warnings in it). Prior to v21 it returned null.

There is a mistake in the client currently, the `unloadwallet` method is
changed in v22 but should have been in v21.

Fix up the types, the client implementation, and the integration tests
to fully test unload.

Add an integration test function that creates, loads, unloads a random
wallet.

As a side note, keep the test names `load_wallet`, `unload_wallet` and
call the other function - this is done so that the tests can be verified
with the `verify` tool.