Retry initial client connections
Previously, `corerpc-node` would try to establish a connection via the
given `Auth` cookie, and would panic if it couldn't setup a initial
connection. Only the rest of the initial setup steps would be retried
with time-delay.

This however was race-y, as the auth cookie might not already be synced
to disk on first startup of a fresh `bitcoind` instance (with fresh data
dir), when we try to connect for the first time. Here, we move the
intial connection logic into the retry loop to simply also retry it
instead of panicking.