backport: bump zip in light of RUSTSEC-2020-0071
This is a backport of #143, done manually but originally authored by
Jose.

--

There's a RUSTSEC advisory on time that should be updated to >=0.2.23.
It gets pulled down from corepc-node with download feature from the zip
dependency, see the cargo tree below:

```
time 0.1.45
└── zip 0.5.13
    └── corepc-node 0.7.0
```

Hence we need to update the zip dependency.

Co-authored-by: Jose Storopoli <jose@storopoli.com>