run-bitcoind: Add err function back in
This got deleted somehow and its still used. I was surprised that
`shellcheck` does not find this, I would have thought a shell linter
would have access to `$PATH` and be able to tell that the function does
not exist. Anywho, add the function back in from a previous release.

Fix: #168