types: Clean up the ScriptPubkey type
The `ScriptPubkey` type is a type that is returned by Core in various
places. It seems like the Core devs put a fair bit of effort into
it (probably because of this) and when the type changed in v22 it was
changed in a way that was completely backwards compatible.

I failed to fully understand this when first adding the type.

With this patch applied there is now only one `ScriptPubkey` type (in
`libr.s`) and also we add a model type.

The `ScriptPubkey` type is unusual in that its fields can be used to
create `TxOut` if tx value is known - which for example in `GetTxOut` it
is.

So here we also remove the v22 `GetTxOut` type since it only existed
because of the perceived difference in `ScriptPubkey` which no longer
exists.

Fix: #166