run-bitcoind: Fix startup of Core nodes
Core v28 introduced changes making node startup fail if bind fails. We
fixed this in `node` in #46 but in `run-bitcoind.sh` we still use
`-port`. This didn't cause any problems because we only pass the arg
to one of the two nodes so port 18445 was available. When we try to
add support for v29 things break.

Jamil was experiencing this problem (I think) while running v28 nodes
but I don't know why.

Use `-bind` instead of `-port`.