integration_test: Pause after node comes up
We are getting intermittent test failures stating that the node is not
available. `node` already includes code that hits the RPC API after the
node is up so this is a little surprising. However it doesn't hurt to
try waiting a little still.

Just sleep for half a second after each node spins up and see if the
intermittent failures go away.