Standardize version format in docs
There is a variation in the way the version number is written in the
docs. Some are preceded by a 0 and some arn't. Core states the versions
up to and including 21 with preceding 0, and for 22 and above without.

Change all version numbers in the text to be v0.17 - v0.21, and v22 -
v29 to match core.