Add missing return fields in v19
Some of the types in v19 had missing return fields. The redefinition of
GetNetworkInfo in v28 was also missing the field.

Add all of the missing fields and update the models, into function
and reexports accordingly.

Redefinitions of the types and into functions are copies of the v17
versions with the missing field added.