Add missing return fields to createmultisig
One of the methods, `createmultisig` was missing a field in v20. It is
missing another field in v23.

Add all of the missing fields and update the models, into function
and reexports accordingly.

Redefinitions of the types and into functions are copies of the v17
versions with the missing field added.