Improve GetPeerInfo
In #310 I got mixed up and added fields to the v17 types `GetPeerInfo`
struct that do not appear in the docs.

This was masked by using `Option`. And not helped by the fact that a
bunch of other `Option`s are used because the docs do not match what
Core returns.

These bugs propagated up through the version numbers. Go over them all
and sort it out.

Note also that we were incorrectly re-exporting `PeerInfo` from v19
even though the `PeerInfo` struct that was actually being used is
found in the version specific file along with `GetPeerInfo` - face
palm.

And finally, remove all the 'vXY and later' docs, this is implied by
the module. These sorts of comments should only be found in `model`.