Fix getaddressinfo
The test was marked as TODO and the fields in the structs do not match
what is returned by the RPC.

Fix all of the fields in v17.

Redefine the types for changes in v18, v29, v22 and v28.  Update the
reexports.

Update the model for to allow for the changes in return types.

Update the error in v18 and reorganise wallet imports.

Update the test using different Address types to check the inner structs
contain expected values.