Add test for listaddressgroupings
`listaddressgroupings` was untested.

Add a test for it that returns a non emply list.

When Serde deserializes from a JSON array (a sequence), it uses 
positional fields and requires the sequence length to match the number 
of struct fields exactly. 

This causes an error when the optional field `label` is not returned.

Change the struct to an enum with 2 fields, one for 2 returned items 
and one for 3.