Test and update getwalletinfo
`getwalletinfo` was implemented for v17 only and untested. There are 
return changes in v18, v19, v21, v23 and v26.  Add the updated structs 
in all versions, update the model and add new  `into_model()`.  Add a 
test with version specific sections to test the new sub-structs.  
Update the types tables and reexports.