Fix verify RPC method tags
During the updating of the RPCs some of them changed from having a
model to not or vice versa. There is also one typo and a duplicate
entry for `enumeratesigners` which was added to `signer` and not removed
from `wallet`.

Fix the tags in verify to match the correct RPC method definitions.