Specify version of dependencies instead of path
I just tried to publish `corepc-client 0.10.0` and got an error from
`cargo`.

  error: all dependencies must have a version specified when publishing.

We recently switched to using `path` instead of versions, seems like
this was wrong.

Seems that the patch sections in the top level manifest are now
redundant but I cannot be bothered to work that out right now.