Depend on bitreq
We just forked `minreq` into this repository to create an HTTP crate
that is specifically designed to be used by Bitcoin projects. As such
we called it `bitreq`.

Depend on the local `bitreq` crate instead of `minreq`. Done directly
in `node` and `jsonrpc`. Indirectly in `client`, including rename of
a million macros that have `minreq` in the name.