Tie electrsd into CI
During import of `electrsd` I just clobbered all CI. Add `electrsd` to
the CI infrastructure in this repo. Use `extra_tests.sh` to mirror the
behaviour in the original repo _excluding_ the no-download test.
Skipping the no-download test because I'm too lazy to set it up, we
don't test no-download in `corepc-node` either.